visual_skills_moe/pyproject.toml
[project]
name = "visual-skills-moe"
version = "0.1.0"
description = "Mixture-of-Experts framework for video understanding with skill routing"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "deepface>=0.0.89",
    "faster-whisper>=1.2.1",
    "gradio>=4.0",
    "litellm>=1.81.5",
    "open-clip-torch>=2.24",
    "opencv-python>=4.8",
    "paddleocr>=2.7",
    "paddlepaddle>=2.6",
    "pandas>=2.0",
    "pyarrow>=15.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "qwen-omni-utils>=0.0.8",
    "soundfile>=0.13.1",
    "tf-keras>=2.16",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.65",
    "transformers>=4.51,<5",
    "ultralytics>=8.1",
]
